Unify clip management under one command
