+++ src/commands/test.js.ignore
@@ -1,14 +1,14 @@
-import { SlashCommandBuilder } from 'discord.js'
+// import { SlashCommandBuilder } from 'discord.js'
 
-import nconf from 'nconf'
-const prefix = nconf.get('COMMAND_PREFIX') || ''
+// import nconf from 'nconf'
+// const prefix = nconf.get('COMMAND_PREFIX') || ''
 
-export let data = new SlashCommandBuilder()
-  .setName(prefix+'test')
-  .setDescription('Test a clip to the soundboard')
-export async function execute(interaction) {
-  await interaction.reply('Bingo!');
-}
+// export let data = new SlashCommandBuilder()
+//   .setName(prefix+'test')
+//   .setDescription('Test a clip to the soundboard')
+// export async function execute(interaction) {
+//   await interaction.reply('Bingo!');
+// }
 
 
 // import { SlashCommandBuilder, ModalBuilder, SelectMenuBuilder, TextInputStyle, TextInputBuilder, ActionRowBuilder, SlashCommandAttachmentOption} from 'discord.js'
